separated  PHP added loading screen
separated big sections of php out of the scan.php page. added two php include statements
block that reads file contents in array is now in readscanresults.php
block that writes a device result to the devicelist is now in writetodevice.php
created form button to scan users network. scansplash.php displays a loading gif while flagging scanCheck.txt to a value of 1 which lets the client pi device know a scan needs to be run.